Added Command Helper and PrimaryTeam command
Command helper will store common methods or variables that we use in
commands, for instance building the team from the arguments.
Added command helper team builder to RandomMember, TeamAdd, and
TeamRemove
Added argument check for commands that take arguments
Added PrimaryTeam command that takes in a team token and boolean and
outputs a list of team members who do or don't have the team set as
their primary team.